+++ IPtProxy.go/IPtProxy.go
@@ -5,6 +5,7 @@ import (
 	snowflakeclient "git.torproject.org/pluggable-transports/snowflake.git/client"
 	"github.com/Yawning/obfs4.git/obfs4proxy"
 	"os"
+	"runtime"
 )
 
 //goland:noinspection GoUnusedConst
@@ -28,9 +29,29 @@ const SnowflakeSocksPort = 52610
 var snowflakeRunning = false
 var obfs4ProxyRunning = false
 
+// Override TOR_PT_STATE_LOCATION, which defaults to "$TMPDIR/pt_state".
+var StateLocation string
+
+func init() {
+	if runtime.GOOS == "android" {
+		StateLocation = "/data/local/tmp"
+	} else {
+		StateLocation = os.Getenv("TMPDIR")
+	}
+
+	StateLocation += "/pt_state"
+}
+
 // Start the Obfs4Proxy.
+//
+// @param logLevel Log level (ERROR/WARN/INFO/DEBUG). Defaults to ERROR if empty string.
+//
+// @param enableLogging Log to TOR_PT_STATE_LOCATION/obfs4proxy.log.
+//
+// @param unsafeLogging Disable the address scrubber.
+//
 //goland:noinspection GoUnusedExportedFunction
-func StartObfs4Proxy() {
+func StartObfs4Proxy(logLevel string, enableLogging, unsafeLogging bool) {
 	if obfs4ProxyRunning {
 		return
 	}
@@ -39,10 +60,27 @@ func StartObfs4Proxy() {
 
 	fixEnv()
 
-	go obfs4proxy.Main()
+	go obfs4proxy.Start(logLevel, &enableLogging, &unsafeLogging)
 }
 
 // Start the Snowflake client.
+//
+// @param ice Comma-separated list of ICE servers.
+//
+// @param url URL of signaling broker.
+//
+// @param front Front domain.
+//
+// @param logFile Name of log file. OPTIONAL
+//
+// @param logToStateDir Resolve the log file relative to Tor's PT state dir.
+//
+// @param keepLocalAddresses Keep local LAN address ICE candidates.
+//
+// @param unsafeLogging Prevent logs from being scrubbed.
+//
+// @param maxPeers Capacity for number of multiplexed WebRTC peers. DEFAULTs to 1 if less than that.
+//
 //goland:noinspection GoUnusedExportedFunction
 func StartSnowflake(ice, url, front, logFile string, logToStateDir, keepLocalAddresses, unsafeLogging bool, maxPeers int) {
 	_, _ = fmt.Fprint(os.Stdout, "Try to start snowflake.", snowflakeRunning)
@@ -55,7 +93,7 @@ func StartSnowflake(ice, url, front, logFile string, logToStateDir, keepLocalAdd
 
 	fixEnv()
 
-	go snowflakeclient.Start(ice, url, front, logFile, logToStateDir, keepLocalAddresses, unsafeLogging, maxPeers)
+	go snowflakeclient.Start(ice, url, front, logFile, &logToStateDir, &keepLocalAddresses, &unsafeLogging, maxPeers)
 }
 
 //goland:noinspection GoUnusedExportedFunction
@@ -75,13 +113,14 @@ func StopSnowflake() {
 // required, or values that we want. have to do this here, since we can only
 // launch this in a thread and the manipulation of environment variables
 // from within an iOS app won't end up in goptlib properly.
+//
+// Note: This might be called multiple times when using different fuctions here,
+// but that doesn't necessarily mean, that the values set are independent each
+// time this is called. It's still the ENVIRONMENT, we're changing here, so there might
+// be race conditions.
 func fixEnv() {
 	_ = os.Setenv("TOR_PT_CLIENT_TRANSPORTS", "meek_lite,obfs2,obfs3,obfs4,scramblesuit,snowflake")
 	_ = os.Setenv("TOR_PT_MANAGED_TRANSPORT_VER", "1")
 
-	tmpdir := os.Getenv("TMPDIR")
-	if tmpdir == "" {
-		os.Exit(1)
-	}
-	_ = os.Setenv("TOR_PT_STATE_LOCATION", tmpdir+"/pt_state")
+	_ = os.Setenv("TOR_PT_STATE_LOCATION", StateLocation)
 }